pyproject.toml
[project]
name = "fastapi-auth"
version = "0.1.0"
description = "auth service in fastapi"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.5",
    "jinja2>=3.1.4",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.2",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.32.1",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
]
